Corrections after review. Index Cursor moved to new file.
