+++ src/pmse_sorted_data_interface.h
@@ -51,10 +51,12 @@ namespace mongo {
 class PmseSortedDataInterface : public SortedDataInterface {
 public:
 
-    PmseSortedDataInterface(StringData ident, const IndexDescriptor* desc, StringData dbpath);
+    PmseSortedDataInterface(StringData ident, const IndexDescriptor* desc,
+                            StringData dbpath);
 
     virtual SortedDataBuilderInterface* getBulkBuilder(OperationContext* txn,
-    bool dupsAllowed) override;
+                                                       bool dupsAllowed)
+                                                                       override;
 
     virtual Status insert(OperationContext* txn, const BSONObj& key,
                           const RecordId& loc, bool dupsAllowed);
@@ -68,8 +70,7 @@ public:
         return Status::OK();
     }
 
-    virtual void fullValidate(OperationContext* txn,
-                              long long* numKeysOut,
+    virtual void fullValidate(OperationContext* txn, long long* numKeysOut,
                               ValidateResults* fullResults) const {
         // TODO: Implement fullValidate
     }
@@ -97,8 +98,6 @@ public:
 
     std::unique_ptr<SortedDataInterface::Cursor> newCursor(
                     OperationContext* txn, bool isForward) const;
-
-
 
 private:
     void moveToNext();
